+++ spike/File/__init__.py
@@ -3,12 +3,21 @@
 
 
 """
-Utilities
+File Utilities
 
 Created by Marc-André on 2011-03-20.
 Copyright (c) 2011 IGBMC. All rights reserved.
 """
-
+__all__ = [
+    'BrukerNMR',
+    'BrukerSMX',
+    'Spinit',
+    'Apex',
+    'Solarix',
+    'Thermo',
+    'mzXML',
+    'csv',
+    'HDF5File']
 
 def main():
     pass