+++ release_notes.md
@@ -1,6 +1,18 @@
 <!-- DO NOT MODIFY this file, it is generated automatically from dev_setup ! -->
 # SPIKE Relase Notes
-#### 0.99.5 - April 2019
+#### 0.99.7 - April 2019
+- first version of Interactive Notebooks:
+  - ProcessFTICR-MS
+  - DisplayFTICR2D
+  - Process1DNMR
+- improved Interactive tools
+- improved FTICR importers to accept `my_expt.d/fid` as well as `my_expt.d`
+- improved peak-picker behavior
+- The BrukerMS importer trys Solarix importer and falls back to Apex importer if it fails
+- improved error messages in FTICR importers
+- cured a bug in Apex.Importxx for a special xml format
+
+#### 0.99.6 - April 2019
 - extended and improved tests - finalized installation through PYPI
 - support for distribution via pip - you can now do `pip instal spike_py` and spike installed globally on your system.
 - still struggling with correct calibration routines for FTICR ! - proceed carefully ! -